+++ bindings/hyperdrivepy/src/hyperdrive_state_methods/long.rs
@@ -1,4 +1,4 @@
 mod close;
 mod max;
 mod open;
-// mod targeted;
+mod targeted;
